requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-ckeditor==6.7.2
django-js-asset==3.0.1
django-meta==2.4.2
django-taggit==6.1.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
